Enum minidump_common::format::MINIDUMP_STREAM_TYPE

#[repr(u32)]
pub enum MINIDUMP_STREAM_TYPE {
    UnusedStream,
    ReservedStream0,
    ReservedStream1,
    ThreadListStream,
    ModuleListStream,
    MemoryListStream,
    ExceptionStream,
    SystemInfoStream,
    ThreadExListStream,
    Memory64ListStream,
    CommentStreamA,
    CommentStreamW,
    HandleDataStream,
    FunctionTable,
    UnloadedModuleListStream,
    MiscInfoStream,
    MemoryInfoListStream,
    ThreadInfoListStream,
    HandleOperationListStream,
    TokenStream,
    JavaScriptDataStream,
    SystemMemoryInfoStream,
    ProcessVmCountersStream,
    IptTraceStream,
    ThreadNamesStream,
    ceStreamNull,
    ceStreamSystemInfo,
    ceStreamException,
    ceStreamModuleList,
    ceStreamProcessList,
    ceStreamThreadList,
    ceStreamThreadContextList,
    ceStreamThreadCallStackList,
    ceStreamMemoryVirtualList,
    ceStreamMemoryPhysicalList,
    ceStreamBucketParameters,
    ceStreamProcessModuleMap,
    ceStreamDiagnosisList,
    LastReservedStream,
    BreakpadInfoStream,
    AssertionInfoStream,
    LinuxCpuInfo,
    LinuxProcStatus,
    LinuxLsbRelease,
    LinuxCmdLine,
    LinuxEnviron,
    LinuxAuxv,
    LinuxMaps,
    LinuxDsoDebug,
    CrashpadInfoStream,
}

The types of known minidump data streams.

Most of these values are derived from the Microsoft enum of the same name, but the values after LastReservedStream are Breakpad and Crashpad extensions.

Variants

UnusedStream

An unused stream directory entry

ReservedStream0

ReservedStream1

ThreadListStream

The list of threads from the process

See MINIDUMP_THREAD.

Microsoft declares a MINIDUMP_THREAD_LIST struct which is the actual format of this stream, but it is a variable-length struct so no matching definition is provided in this crate.

ModuleListStream

The list of executable modules from the process

See MINIDUMP_MODULE.

Microsoft declares a MINIDUMP_MODULE_LIST struct which is the actual format of this stream, but it is a variable-length struct so no matching definition is provided in this crate.

MemoryListStream

The list of memory regions from the process contained within this dump

See MINIDUMP_MEMORY_DESCRIPTOR.

Microsoft declares a MINIDUMP_MEMORY_LIST struct which is the actual format of this stream, but it is a variable-length struct so no matching definition is provided in this crate.

ExceptionStream

Information about the exception that caused the process to exit

See MINIDUMP_EXCEPTION_STREAM.

SystemInfoStream

System information

See MINIDUMP_SYSTEM_INFO.

ThreadExListStream

Memory64ListStream

CommentStreamA

CommentStreamW

HandleDataStream

FunctionTable

UnloadedModuleListStream

MiscInfoStream

Miscellaneous process and system information

See ['MINIDUMP_MISC_INFO'][struct.MINIDUMP_MISC_INFO.html).

MemoryInfoListStream

Information about memory regions from the process

See ['MINIDUMP_MEMORY_INFO_LIST'][struct.MINIDUMP_MEMORY_INFO_LIST.html).

ThreadInfoListStream

HandleOperationListStream

TokenStream

JavaScriptDataStream

SystemMemoryInfoStream

ProcessVmCountersStream

IptTraceStream

ThreadNamesStream

ceStreamNull

ceStreamSystemInfo

ceStreamException

ceStreamModuleList

ceStreamProcessList

ceStreamThreadList

ceStreamThreadContextList

ceStreamThreadCallStackList

ceStreamMemoryVirtualList

ceStreamMemoryPhysicalList

ceStreamBucketParameters

ceStreamProcessModuleMap

ceStreamDiagnosisList

LastReservedStream

BreakpadInfoStream

Additional process information (Breakpad extension)

See ['MINIDUMP_BREAKPAD_INFO'][struct.MINIDUMP_BREAKPAD_INFO.html).

AssertionInfoStream

Assertion information (Breakpad extension)

See ['MINIDUMP_ASSERTION_INFO'][struct.MINIDUMP_ASSERTION_INFO.html).

LinuxCpuInfo

The contents of /proc/cpuinfo from a Linux system

LinuxProcStatus

The contents of /proc/self/status from a Linux system

LinuxLsbRelease

The contents of /etc/lsb-release from a Linux system

LinuxCmdLine

The contents of /proc/self/cmdline from a Linux system

LinuxEnviron

The contents of /proc/self/environ from a Linux system

LinuxAuxv

The contents of /proc/self/auxv from a Linux system

LinuxMaps

The contents of /proc/self/maps from a Linux system

LinuxDsoDebug

Information from the Linux dynamic linker useful for writing core dumps

See 'DSO_DEBUG_64' and 'DSO_DEBUG_32'.

CrashpadInfoStream