Trait minidump_common::traits::IntoRangeMapSafe

pub trait IntoRangeMapSafe<V>: IntoIterator<Item = (Range<u64>, V)> + Sized where
    V: Clone + Debug + Eq, {
    fn into_rangemap_safe(self) -> RangeMap<u64, V> { ... }
}

This trait exists to allow creating RangeMaps from possibly-overlapping input data.

The RangeMap struct will panic if you attempt to initialize it with overlapping data, and we deal with many sources of untrusted input data that could run afoul of this. Upstream issue

Provided methods

fn into_rangemap_safe(self) -> RangeMap<u64, V>

Loading content...